packages/fireflies-nodes/pyproject.toml
[project]
name = "aceteam-fireflies-nodes"
version = "0.1.0"
description = "Fireflies.ai transcript nodes for AceTeam: list, fetch, and search meeting transcripts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "aceteam-workflow-engine>=2.0.0rc5",
    "httpx>=0.28.0",
]

[project.entry-points."aceteam.nodes"]
fireflies = "fireflies_nodes:register_nodes"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fireflies_nodes"]
